Consolidating two event classes
